Complete overhaul of file structure
This has been done for two reasons. The first is to situate all the information for each page in the same folder, while still keeping a clear seperation between content and markup. The second is to allow for easy integration with a CMS to enable non-technical people to edit the content.

Each page now has its own folder. All pages have a .njk file and a .yaml file. The .njk file provides markup (and can still extend and include shared components from _includes), while the .yaml file holds the content.