Added info about specific concerts
Concerts are specified with a time, a name, and a list of locations. Each location has a name and a list of bands playing at that location.

Multiple locations and multiple bands at each location must be supported because the starting and ending march requires it.

The page for each band displays the concerts that band has, using computed data.