Revamped JS setup
The inline JS pipeline is now gone. Instead you put all js files in src/js/, and add them to a page with the js block.

This means JS is no longer minified, and no longer concatenated into a single file. This could lead to bad performance, which has to be investigated. However, there is going to be very little JS overall, so there might not be a problem. In any case I'd like to see if 11ty's custom file extension handler feature gets released before I look into this further.